+++ src/js/page/terminal/commands/__all__.mjs
@@ -16,6 +16,7 @@ import cmdPrint from './print';
 import cmdQuit from './quit';
 import cmdToggleTerm from './toggle_term';
 import cmdInput from './input';
+import cmdRun from './run';
 import type VMachine from '@trash/vmachine';
 
 export default function (vm: VMachine) {
@@ -34,4 +35,5 @@ export default function (vm: VMachine) {
   vm.registerCommand('dis', cmdDis());
   vm.registerCommand('genfile', cmdGenFile());
   vm.registerCommand('input', cmdInput());
+  vm.registerCommand('run', cmdRun());
 }