+++ src/js/page/terminal/commands/toggle_term.mjs
@@ -4,11 +4,16 @@
 
 // $FlowIgnore
 import i18n from 'i18next';
-import type {CMDDef} from '@trash/interpreter';
+import {ARG} from '@trash/constants';
+import type {CMDCallbackArgs, CMDDef} from '@trash/interpreter';
 import type Terminal from '@terminal/terminal';
 
-async function cmdToggleTerm(this: Terminal): Promise<> {
-  return this.doToggle();
+async function cmdToggleTerm(this: Terminal, kwargs: CMDCallbackArgs): Promise<> {
+  let force_show;
+  if (typeof kwargs.force !== 'undefined') {
+    force_show = kwargs.force === 'show';
+  }
+  return this.doToggle(force_show);
 }
 
 export default function (): Partial<CMDDef> {
@@ -16,5 +21,9 @@ export default function (): Partial<CMDDef> {
     definition: i18n.t('cmdToggleTerm.definition', 'Toggle terminal visibility'),
     callback: cmdToggleTerm,
     detail: i18n.t('cmdToggleTerm.detail', 'Toggle terminal visibility'),
+    args: [
+      [ARG.String, ['f', 'force'], false, i18n.t('cmdToggleTerm.args.force', 'Force show/hide'), undefined, ['show', 'hide']],
+    ],
+    example: "-s true",
   };
 }